genai-recommendations/requirements.txt
pandas
seaborn
matplotlib
python-dotenv
tiktoken
langchain
langchain-google-genai
streamlit
faiss-cpu
sentence-transformers
langchain_community